--- src/__tests__/fixtures/pagination.ts
@@ -1,8 +1,0 @@
-import { Pagination } from '../../types';
-
-const defaultPagination: Pagination = {
-  page: 0,
-  pageSize: 100,
-};
-
-export default defaultPagination;
