+++ src/infra/sequelize/queries/project/lister/requêtes/listerProjetsAcheteurObligé.ts
@@ -1,7 +1,7 @@
 import { getProjectAppelOffre } from '@config/queryProjectAO.config';
 import { ListerProjets } from '@modules/project/queries';
 import { Project } from '@infra/sequelize/projectionsNext';
-import { makePaginatedList, paginate } from '../../../../../../helpers/paginate';
+import { makePaginatedList, mapToOffsetAndLimit } from '../../../pagination';
 import { mapToFindOptions } from '../../helpers/mapToFindOptions';
 import { Op } from 'sequelize';
 
@@ -37,7 +37,7 @@ export const listerProjetsPourAcheteurObligé: ListerProjets = async ({ paginati
       ...findOptions?.where,
       notifiedOn: { [Op.gt]: 0 },
     },
-    ...paginate(pagination),
+    ...mapToOffsetAndLimit(pagination),
     attributes,
   });
