+++ src/modules/notificationCandidats/queries/GetPendingCandidateInvitations.ts
@@ -1,5 +1,5 @@
 import { ResultAsync } from '@core/utils';
-import { PaginatedList, Pagination } from '../../../types';
+import { PaginatedList, Pagination } from '@modules/pagination';
 import { InfraNotAvailableError } from '../../shared';
 
 export type PendingCandidateInvitationDTO = {
