+++ src/views/pages/InvitationsCandidatsEnAttentePage.tsx
@@ -16,7 +16,7 @@ import { PendingCandidateInvitationDTO } from '@modules/notificationCandidats';
 import routes from '@routes';
 import { Request } from 'express';
 import React from 'react';
-import { PaginatedList } from '../../types';
+import { PaginatedList } from '@modules/pagination';
 import { afficherDateAvecHeure, hydrateOnClient } from '../helpers';
 
 interface InvitationsCandidatsEnAttenteProps {
