+++ src/views/pages/ListeProjetsPage.tsx
@@ -2,7 +2,7 @@ import { Request } from 'express';
 import querystring from 'querystring';
 import React, { useState } from 'react';
 import { AppelOffre, Famille, Periode } from '@entities';
-import { PaginatedList } from '../../types';
+import { PaginatedList } from '@modules/pagination';
 
 import {
   ProjectList,
