+++ src/views/pages/ProjetsÀRéclamerPage.tsx
@@ -1,7 +1,7 @@
 import { Request } from 'express';
 import React, { useState } from 'react';
 import { AppelOffre, Famille, Periode, Project } from '@entities';
-import { PaginatedList } from '../../types';
+import { PaginatedList } from '@modules/pagination';
 import {
   BarreDeRecherche,
   Dropdown,
