+++ src/views/pages/modificationRequestListPage/ModificationRequestList.tsx
@@ -2,7 +2,7 @@ import { Request } from 'express';
 import React, { ChangeEvent, useState } from 'react';
 import { AppelOffre } from '@entities';
 import { ModificationRequestListItemDTO } from '@modules/modificationRequest';
-import { PaginatedList } from '../../../types';
+import { PaginatedList } from '@modules/pagination';
 import {
   RequestList,
   LegacyPageTemplate,
